refactor: rename tf_phasespace to tensorflow
